Handling amount and rate in the user balance class

+++ CryptoTrading.Logic/Options/CryptoTradingOptions.cs
@@ -2,6 +2,6 @@
 {
     public class CryptoTradingOptions
     {
-        public int AmountInUsdt { get; set; }
+        public decimal AmountInUsdt { get; set; }
     }
 }
